+++ saleor/static/dashboard-next/components/AssignProductDialog/AssignProductDialog.tsx
@@ -15,7 +15,6 @@ import * as React from "react";
 import ConfirmButton, {
   ConfirmButtonTransitionState
 } from "../../components/ConfirmButton/ConfirmButton";
-import Debounce from "../../components/Debounce";
 import Form from "../../components/Form";
 import FormSpacer from "../../components/FormSpacer";
 import TableCellAvatar from "../../components/TableCellAvatar";
@@ -84,29 +83,25 @@ const AssignProductDialog = withStyles(styles, {
           <>
             <DialogTitle>{i18n.t("Assign Product")}</DialogTitle>
             <DialogContent className={classes.overflow}>
-              <Debounce debounceFn={onFetch}>
-                {fetch => (
-                  <TextField
-                    name="query"
-                    value={data.query}
-                    onChange={event => change(event, () => fetch(data.query))}
-                    label={i18n.t("Search Products", {
-                      context: "product search input label"
-                    })}
-                    placeholder={i18n.t(
-                      "Search by product name, attribute, product type etc...",
-                      {
-                        context: "product search input placeholder"
-                      }
-                    )}
-                    fullWidth
-                    InputProps={{
-                      autoComplete: "off",
-                      endAdornment: loading && <CircularProgress size={16} />
-                    }}
-                  />
+              <TextField
+                name="query"
+                value={data.query}
+                onChange={event => change(event, () => onFetch(data.query))}
+                label={i18n.t("Search Products", {
+                  context: "product search input label"
+                })}
+                placeholder={i18n.t(
+                  "Search by product name, attribute, product type etc...",
+                  {
+                    context: "product search input placeholder"
+                  }
                 )}
-              </Debounce>
+                fullWidth
+                InputProps={{
+                  autoComplete: "off",
+                  endAdornment: loading && <CircularProgress size={16} />
+                }}
+              />
               <FormSpacer />
               <Table>
                 <TableBody>