+++ saleor/static/dashboard-next/containers/SearchCollections/types/SearchCollections.ts
@@ -27,5 +27,7 @@ export interface SearchCollections {
 }
 
 export interface SearchCollectionsVariables {
+  after?: string | null;
+  first: number;
   query: string;
 }