+++ saleor/static/dashboard-next/containers/SearchProducts/types/SearchProducts.ts
@@ -14,8 +14,8 @@ export interface SearchProducts_products_edges_node_thumbnail {
 export interface SearchProducts_products_edges_node {
   __typename: "Product";
   id: string;
-  thumbnail: SearchProducts_products_edges_node_thumbnail | null;
   name: string;
+  thumbnail: SearchProducts_products_edges_node_thumbnail | null;
 }
 
 export interface SearchProducts_products_edges {
@@ -33,5 +33,7 @@ export interface SearchProducts {
 }
 
 export interface SearchProductsVariables {
+  after?: string | null;
+  first: number;
   query: string;
 }