+++ saleor/static/dashboard-next/orders/fixtures.ts
@@ -1,3 +1,4 @@
+import { SearchCustomers_customers_edges_node } from "../containers/SearchCustomers/types/SearchCustomers";
 import { transformOrderStatus, transformPaymentStatus } from "../misc";
 import {
   FulfillmentStatus,
@@ -8,9 +9,8 @@ import {
 } from "../types/globalTypes";
 import { OrderDetails_order } from "./types/OrderDetails";
 import { OrderList_orders_edges_node } from "./types/OrderList";
-import { UserSearch_customers_edges_node } from "./types/UserSearch";
 
-export const clients: UserSearch_customers_edges_node[] = [
+export const clients: SearchCustomers_customers_edges_node[] = [
   {
     __typename: "User" as "User",
     email: "test.client1@example.com",