+++ saleor/static/dashboard-next/products/components/ProductListCard/ProductListCard.tsx
@@ -5,44 +5,53 @@ import * as React from "react";
 
 import { CategoryDetails_category_products_edges_node } from "../../../categories/types/CategoryDetails";
 import Container from "../../../components/Container";
+import { FilterContentSubmitData } from "../../../components/Filter";
 import PageHeader from "../../../components/PageHeader";
 import ProductList from "../../../components/ProductList";
 import { Filter } from "../../../components/TableFilter";
 import i18n from "../../../i18n";
 import { ListActions, PageListProps } from "../../../types";
-import ProductListFilter, { ProductListFilterTabs } from "../ProductListFilter";
+import ProductListFilter from "../ProductListFilter";
 
-interface ProductListCardProps extends PageListProps, ListActions {
-  currentTab: ProductListFilterTabs;
+export interface ProductListCardProps extends PageListProps, ListActions {
+  currencySymbol: string;
+  currentTab: number;
   filtersList: Filter[];
+  initialSearch: string;
   products: CategoryDetails_category_products_edges_node[];
   onAllProducts: () => void;
-  onAvailable: () => void;
-  onCustomFilter: () => void;
-  onOfStock: () => void;
+  onSearchChange: (value: string) => void;
+  onFilterAdd: (filter: FilterContentSubmitData) => void;
+  onFilterDelete: () => void;
+  onFilterSave: () => void;
+  onTabChange: (tab: number) => void;
 }
 
 export const ProductListCard: React.StatelessComponent<
   ProductListCardProps
 > = ({
+  currencySymbol,
   currentTab,
   disabled,
   filtersList,
+  initialSearch,
   pageInfo,
   products,
   onAdd,
   onAllProducts,
-  onAvailable,
-  onCustomFilter,
   onNextPage,
-  onOfStock,
   onPreviousPage,
   onRowClick,
   isChecked,
   selected,
   toggle,
   toggleAll,
-  toolbar
+  toolbar,
+  onSearchChange,
+  onFilterAdd,
+  onFilterSave,
+  onTabChange,
+  onFilterDelete
 }) => (
   <Container>
     <PageHeader title={i18n.t("Products")}>
@@ -52,12 +61,16 @@ export const ProductListCard: React.StatelessComponent<
     </PageHeader>
     <Card>
       <ProductListFilter
+        currencySymbol={currencySymbol}
         currentTab={currentTab}
         filtersList={filtersList}
-        onAvailable={onAvailable}
+        initialSearch={initialSearch}
         onAllProducts={onAllProducts}
-        onOfStock={onOfStock}
-        onCustomFilter={onCustomFilter}
+        onSearchChange={onSearchChange}
+        onFilterAdd={onFilterAdd}
+        onFilterSave={onFilterSave}
+        onTabChange={onTabChange}
+        onFilterDelete={onFilterDelete}
       />
       <ProductList
         products={products}